+++ requests/config.py
@@ -54,3 +54,5 @@ class Settings(object):
         return object.__getattribute__(self, key)
 
 settings = Settings()
+settings.base_headers = {'User-Agent': 'python-requests.org'}
+settings.accept_gzip = True